Fixed Options
Options were broken with the options tab update.

This commit was mainly focused at fixing the zoom which a lot of other things relied on.

There's some things that are still broken but will have to be fixed in the future when support to the new settings interface is added.