Update bank.simba
Bank searching had several problems that showed up in my use case.

The old way it was done, if an item was search previously and Enter was pressed it would not be able to search again because the query would be closed but the bank was still in search mode.

It now should be able to search multiple items without issues.