Master fixes (#14)
* Bank.Hover fix

* Update inventory.simba

Removed Self.IsOpen from TRSInventory.Open because GameTabs.Open already checks for it.
No point in having a double check I think.

* Update inventory.simba

Sets a result for Inventory.Use

* Update inventory.simba

Better result? Just because we find the items doesn't necessarily mean we will click them I guess? Not a big deal anyway.

* remoteinput update

* remoteinput update

* Update minimap.simba

* Update prayer.simba

Updates minimap prayer to handle the new active colors

* Update interface.simba

ButtonColors should include EnabledButtonColors

* Update options.simba

* Update mm2ms.simba

* Update itemfinder.simba

fix for linux

* Update chat.simba

* Update inventory.simba

* Update equipment.simba

I just think this makes more sense and is more in line with the inventory file.

* Update equipment.simba

* Update inventory.simba

function to check for items without extra parameters other than the item name/id

* Update chat.simba

Split AnswerQuery into FindQuery and AnswerQuery.
This is useful in case if you want to check if a Query is open but not necessarily want to interact with it.
This solved me some issues with bank searching for example.

* Update bank.simba

Bank searching had several problems that showed up in my use case.

The old way it was done, if an item was search previously and Enter was pressed it would not be able to search again because the query would be closed but the bank was still in search mode.

It now should be able to search multiple items without issues.

* Update bank.simba

This fixes a issue I was having with Simba getting stuck trying to open inventory.

I'm not really sure how to recreate it but it happened while using the old bank search

* Update chat.simba

* Update chat.simba

just a comment

* Update chatbuttons.simba

Full chat buttons functionality

* Update interface.simba

I think we should have the ability to choose which mouse button to click with. I required this for my chat buttons

* Update osr.simba

reordered chooseoptions so it's included before chatbuttons.