Experimental: Adaptive walking,  mouse-ahead for mainscreen
Adaptive walking can be triggered with
`RSW.AdaptiveWalk := True`
`RSW.FancyMouse` now also works for mainscreen, might not be as pretty.

Adpative walking mixes mainscreen and minimap walking. The first few steps can happen on the mainscreen, and same for the last few steps. This is perhaps currently a little too consistent. But it's an experimental feature. MainScreen walking in itself seems at best questionable, path fails to advance more often it seems.