perf: reworked some inventory methods for performance
Some inventory methods were using `TRSInventory.FindItems()` in them which in turn uses `ItemFinder.Find()` with `MaxToFind = -1`.

This is not required for any of the `Contains` methods since we only have to find one and exit.

This was not tested but it's obviously going to be faster since there's less needless searching.

I've also reworked `TRSInventory.Count()` to be at least 50% faster, in some cases 75%.
The previous version also failed consistently with inventory background set to transparent in resizable, the new version, could still fail in theory but from my testing it never did.