TRSWalker1500 backport (#66)
* refactor: added `Contains` methods to equipment and inventory.

I've also left a comment on the PR after the last commit because I forgot to mention TRSBank.ContainsItem().

Anyway, this has the same reasoning. This is optional but I would like to bring the SRL1400 API closer to SRL1500 API.

Old methods are kept and marked with a comment as deprecated so there's no downside to adding this.

This also brings documentation to TRSInventory which is nice :)

* fix: added deprecated warnings

also added more docs to inventory

* feat: XPBar rework

Current XPBar has a lot of issues:

- Because it's position and size are never permanent, there are situation where they are set wrong and then reading it fails.

- It also doesn't work with progress bar enabled which should be optional.

- Large uptexts in fixed mode can make SRL think the xpbar is not open.

This rework fixes all those. It also allows reading when is partially covered, which can be seen as a good or bad thing but I think it's fine and should be handled by the user.

* fix: adding some utilities for walker

They might also be useful to have for other things.

* feat: TBoxArray.Pack()

credits go to Olly. I've just copy pasted it from Simba1500 source and made some tweaks for it to compile.

Basically, moves TBox coordinates in a TBoxArray so they ocupy the least area possible.