fix: TRSBank.IsOpen()
The current way the bank is open is , it's not always reliable.
The current way basically relies on the button amount to be exactly what is expected, if it's not, for SRL, the Bank is closed.
There's several reasons buttons can be covered sometimes which makes this very unreliable at times.

For example, if a user has the incinerator enabled and the mouse is in that position, a tooltip will show up that covers the buttons. And yes, this could be considered user fault for not having things setup properly but the incinerator on is actually the default and this is just one example of many that can make SRL think the bank is not open.

My proposed way should be much more reliable and I've been testing for over 2 weeks and seems to work fine.