feat: add TRSBankItem
This commit adds a new record called `TRSBankItem`.
```pascal
TRSBankItem = record
    Item: TRSItem;
    Quantity: Int32;
    Noted: Boolean;

    Tab: Int32;
    Scroll: Int32;

    IsSetup: Boolean;
  end;
```

It's purpose long term would be to completely replace `TRSBankWithdrawItem` and `TRSBankDepositItem`. Due to how widely used this are I choose to keep them for now. Marking methods that use them as deprecated could be studied.

I also made some changes to the core `TRSBank` by adding the currently global constants of the file directly to it as class constants.
The idea would be for this to also replace the current constants in the future completely but I also left them for the same reason I left `TRSBankWithdrawItem` and `TRSBankDepositItem`. It's widely used.

`TRSBank.WithdrawHelper()` was edited to cache the last withdrawn custom quantity. This is something I've been doing for a long time in WaspLib and has been working great in my opinion.
Reading up text while the bank is open has a decent probability to fail, I don't know the odds but it's higher than normal, this mitigates the issue by a lot, by keeping the last withdrawn value cached in a new `class var` added to `TRSBank` called `CachedQuantity`.

Lastly, a new `TRSBank.WithdrawItem()` and `TRSBank.DepositItem()` that use the new `TRSBankItems` were added.